app/chatbot/backend/requirements.txt
# Framework web
fastapi==0.100.0
uvicorn[standard]==0.23.0

# Banco de dados (opcional, dependendo da implementação)
databases[postgresql]==0.6.0
psycopg2-binary==2.9.9

# Middleware CORS
fastapi-cors==0.0.6

# Outros utilitários
pydantic==2.2.1
requests==2.31.0
python-multipart==0.0.9
pdf2image==1.17.0

# Logging
loguru==0.7.0

# Variáveis de ambiente
python-dotenv==1.0.0

#Images
pillow==10.4.0

# LLM Models
google-generativeai==0.7.2
transformers==4.45.2

# Manipulação de imagens e OCR
Pillow==10.4.0
pytesseract==0.3.13

# Reconhecimento de fala
speechrecognition==3.10.4
pydub==0.25.1

# Necessário para transformers

tensorflow==2.16.1
tf-keras
torch==2.4.1